+++ Core/GraphicsUtils.h
@@ -10,8 +10,8 @@ namespace Math
 		inline float Radians(float degrees);
 		inline float Degrees(float radians);
 
-		Math::Matrix RotationXY(float angle, int scale = 4);
-		Math::Matrix RotationYZ(float angle, int scale = 4);
-		Math::Matrix RotationXZ(float angle, int scale = 4);
+		Math::Matrix RotationXY(float angle, size_t scale = 4);
+		Math::Matrix RotationYZ(float angle, size_t scale = 4);
+		Math::Matrix RotationXZ(float angle, size_t scale = 4);
 	};
 }
